reworked output_writer module
